further enhancements on 3437, remove hphelper
